+++ sdk/eventgrid/azure-eventgrid/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.0.0b5 (Unreleased)
 
+  **Breaking Changes**
+  - `EventGridSharedAccessSignatureCredential` is deprecated in favor of `AzureSasCredential`.
 
 ## 2.0.0b4 (2020-11-11)